get docs building again
Docs were not building on Documenter 1 due to the repo link, but this doesn't seem needed, since the GitHub link in the rendered docs points to the right location.

This removes the accidentally committed docs/build, and updates .gitignore to ensure that won't happen again.

Updated the "Manual test" code and made sure it is working.